pyproject.toml
[tool.poetry]
name = "cgcg"
packages = [ {include = "src"} ]
version = "0.1.0"
description = ""
authors = ["Natalia Koliou <natalykoliou@yahoo.gr>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
torch = "^2.3.1"
numpy = "^1.26.4"
matplotlib = "3.8.2"
pyyaml = "^6.0.1"
pygame = "^2.6.0"
tqdm = "^4.66.1"
psutil = "^5.9.4"
open_spiel = "^1.5"
ruamel-yaml = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
learner = "src.learner:main"
simulator = "src.simulator:main"
evaluator = "src.evaluator:main"
designer = "src.designer:main"
temp = "src.temp:main"